requirements.txt
Flask
pyodbc
python-dotenv
numpy
